Removing alpha from add/sub/mul/div operations
This update was already implemented in the master branch a while ago,
but seems to have been accidentally forgotten on the condensed branch.

The whole reason for removing alpha from the arithmetic operations is
that it produces undesired effects. E.g., Color(255,255,255) -
Color(255,0,0) would produce Color(0,255,255,0) because the lefthand
side's alpha (255) was being subtracted by the righthand side's alpha
(255).